+++ tests/files/src/0888_void_expression.php
@@ -1,0 +1,6 @@
+<?php
+function do_echo(int $i): void {
+    echo $i;
+}
+$x = [do_echo(1) => do_echo(2)];
+echo (string)do_echo(3);